basic App tests - working fetch mocks
