+++ src/App.test.tsx
@@ -1,0 +1,26 @@
+import { cleanup, screen } from '@testing-library/react';
+import { renderRoute } from './testUtils';
+
+afterEach(cleanup);
+
+const user = '{"email":"u1@gmail.com","firstname":"Alex","lastname":"Mayers"}';
+const rewards = `[{"id": 6,"from": ${user},"to": ${user},"reward": 5,"datetime": "2021-06-19T11:15:14.856Z","message": "Thanks"}]`;
+
+describe('App', () => {
+  test('renders Spinner when loading', () => {
+    renderRoute('/');
+    const spinner = screen.getByTestId('spinner');
+    expect(spinner).toBeInTheDocument();
+  });
+
+  test('renders app with Add Reward button on load', async () => {
+    jest
+      .spyOn(global, 'fetch')
+      .mockResolvedValueOnce(Promise.resolve(new Response(rewards)))
+      .mockResolvedValueOnce(Promise.resolve(new Response(user)));
+
+    renderRoute('/');
+    const addRewardButton = await screen.findByRole('button', { name: /add reward/i });
+    expect(addRewardButton).toBeInTheDocument();
+  });
+});
